Add support for running upgrade script before updating dependencies (#28)
## Motivation

With Comet v8 we will remove the `@comet/blocks-api` and
`@comet/blocks-admin` packages. These packages have to be removed from
the corresponding package.json files before updating the dependencies,
as there would be an install error due to the missing packages
otherwise.

## Solution

Introduce the concept of stages (could also be called hooks or similar)
that allows an upgrade script to define when it should be executed. The
stage can be defined by exporting a magic constant `stage` (similar to
what Next does with `dynamic`). If not supplied, the default stage will
be `after-install`.

## Example

```ts
// Script should be run before installing dependencies
export const stage = "before-install";

export default async function upgradeScript() {
    // Perform some actions before installing dependencies
}
```